MINOR: [Docs][Python] Fix return type in docstring for Array.slice (#44134)
### Rationale for this change

Currently the docstring for Array.slice says it returns a RecordBatch. I don't see how this is possible with the existing code. My guess is that this was a copy-and-paste error back when the Array and RecordBatch slice impls were added in https://issues.apache.org/jira/browse/ARROW-547.

### What changes are included in this PR?

Just a docstring change. I copied the language from the take method so things are consistent.

### Are these changes tested?

No.

### Are there any user-facing changes?

Just docs.

Authored-by: Bryce Mecum <petridish@gmail.com>
Signed-off-by: Bryce Mecum <petridish@gmail.com>