MINOR: [Docs][Python] Document example for pc.index (#44537)
### Rationale for this change

In another project, I've been referring to the PyArrow Compute documentation quite a lot, and I believe having more documentation here would be helpful. 

I added an example for Index, also to highlight the behavior of `start` and the -1 return if the value isn't found. 

### What changes are included in this PR?

Adding an example that should appear on the [pyarrow.compute.index](https://arrow.apache.org/docs/python/generated/pyarrow.compute.index.html#) documentation page. 

### Are these changes tested?

I ran this code using the latest version of PyArrow and ensured the outputs were as expected. I did not test building the documentation and rendering, although I would like to start doing this and learn more about this process for future documentation purposes. 

### Are there any user-facing changes?

A new example would be available on the website. 

Authored-by: Tyler White <50381805+IndexSeek@users.noreply.github.com>
Signed-off-by: Sutou Kouhei <kou@clear-code.com>