GH-43808: [C++] skip `-0117` in StrptimeZoneOffset for old glibc (#44621)
### Rationale for this change

Enable tests for libarrow in conda-forge: https://github.com/apache/arrow/issues/35587

### What changes are included in this PR?

old glibc does not actually support timezones like `-0117` (used in `StrptimeZoneOffset` test). The exact lower bound for glibc is hard for me to determine; I know that it passes with 2.28 and that it fails with 2.17. Anything in between is an open question. I went with the conservative option here.

### Are these changes tested?

Tested in https://github.com/conda-forge/arrow-cpp-feedstock/pull/1058

### Are there any user-facing changes?

* GitHub Issue: #43808

Lead-authored-by: H. Vetinari <h.vetinari@gmx.com>
Co-authored-by: Sutou Kouhei <kou@cozmixng.org>
Signed-off-by: Sutou Kouhei <kou@clear-code.com>