MINOR: [Docs][Python] Fix a few minor typos (#44756)
### Rationale for this change

I noticed a few typos in the documentation and a single typo in the pyproject file. This change corrects these typos. 

### What changes are included in this PR?

The following spelling corrections have been applied:

fo -> to
scenerios -> scenarios
wich -> which
suiteable -> suitable
avaiable -> available (this is in the pyproject.toml file regarding a NumPy comment for older Python versions)

### Are these changes tested?

No executable code is modified with these changes.

### Are there any user-facing changes?

Only the documentation is being updated.

Authored-by: Tyler White <50381805+IndexSeek@users.noreply.github.com>
Signed-off-by: Sutou Kouhei <kou@clear-code.com>