GH-44952: [C++][Python] Add Hyperbolic Trig functions (#44630)
### Rationale for this change

Hyperbolic trigonometric functions are a common transformation for dealing with skewed data. And they are built into the core C++ libraries so require minimal change.

### What changes are included in this PR?

Adding `a?(sin|cos|tan)h` to the base C++ library and substrait and tests for these functions in pyarrow.

### Are these changes tested?

Yes, in the same style as the trigonometric functions.

### Are there any user-facing changes?

Yes. Additional compute functions are added.
* GitHub Issue: #44952

Authored-by: Kevin H Wilson <khwilson@gmail.com>
Signed-off-by: Felipe Oliveira Carvalho <felipekde@gmail.com>