GH-45034: [C++] Remove Parquet requirement from Arrow Acero and from Arrow Dataset when not necessary (#45035)
### Rationale for this change

Currently we are defining Parquet as a requirement for Acero and for Dataset on `ArrowAceroConfig.cmake.in` and `ArrowDatasetConfig.cmake.in` respectively.

Parquet should never be necessary for Acero and could be necessary for Dataset only if it has been built with Parquet support.

### What changes are included in this PR?

Remove Parquet from `ArrowAceroConfig.cmake.in` and add it as optional dependency for `ArrowDatasetConfig.cmake.in`

### Are these changes tested?

I've tested with a really minimal build of Arrow and Pyarrow. I don't think we want to add a build without Parquet but with Dataset and/or Acero at CI just to test this scenario.
Logs for the local build without Parquet:
```
-- Found Cython version: 3.0.11
-- Arrow version: 19.0.0
-- Found the Arrow shared library: /home/raulcd/code/dist/lib/libarrow.so.1900.0.0
-- Found the Arrow import library: ARROW_IMPORT_LIB-NOTFOUND
-- Found the Arrow static library: 
-- Building PyArrow with Dataset
-- ArrowAcero version: 19.0.0
-- Found the ArrowAcero shared library: /home/raulcd/code/dist/lib/libarrow_acero.so.1900.0.0
-- Found the ArrowAcero import library: ARROW_ACERO_IMPORT_LIB-NOTFOUND
-- Found the ArrowAcero static library: 
-- ArrowDataset version: 19.0.0
-- Found the ArrowDataset shared library: /home/raulcd/code/dist/lib/libarrow_dataset.so.1900.0.0
-- Found the ArrowDataset import library: ARROW_DATASET_IMPORT_LIB-NOTFOUND
-- Found the ArrowDataset static library: 
-- Building PyArrow with Acero
-- Parquet Encryption is NOT Enabled
```

The `IMPORT_LIB-NOTFOUND` has always been present and it's not new.

### Are there any user-facing changes?

No as this should have already been the case.
* GitHub Issue: #45034

Authored-by: Raúl Cumplido <raulcumplido@gmail.com>
Signed-off-by: Sutou Kouhei <kou@clear-code.com>