GH-45204: [Integration][Archery] Remove skips for nanoarrow IPC compression ZSTD/uncompressible golden files (#45205)
### Rationale for this change

After apache/arrow-nanoarrow#693 , ZSTD compression is now supported in the nanoarrow IPC reader. The list of skips lives in archery, though, and I'd like those checks to run (here and on our own CI!).

### What changes are included in this PR?

The line skipping compression checks for nanoarrow IPC were modified to only skip lz4 (which is not yet implemented).

### Are these changes tested?

Yes, this code runs as part of the integration CI job. The skipped tester is not run in the Arrow repo, though (because of the "target implementations", which correctly doesn't include nanoarrow here); however, the changes are tested in https://github.com/apache/arrow-nanoarrow/pull/704 .

(That PR will need to merge before this one because this PR updates the nanoarrow build script in a way will cause the integration job to fail before that PR is merged)

### Are there any user-facing changes?

No!
* GitHub Issue: #45204

Lead-authored-by: Dewey Dunnington <dewey@wherobots.com>
Co-authored-by: Dewey Dunnington <dewey@fishandwhistle.net>
Signed-off-by: Dewey Dunnington <dewey@fishandwhistle.net>