MINOR: [Dev] Minor edits and slight reorganization of usage_question issue template (#45450)
### Rationale for this change

A conversation on Zulip reminded me we have this issue template and, after reviewing, I thought the text could be improved a bit. My main goal was to make it more clear that the user@ list is the place users may want to go with usage questions.

### Are these changes tested?

Just by pulling the markdown out and ensuring it renders locally. They do.

### Are there any user-facing changes?

Updated issue template.

Authored-by: Bryce Mecum <petridish@gmail.com>
Signed-off-by: Bryce Mecum <petridish@gmail.com>