GH-45266: [C++][Acero] Fix the running tasks count of Scheduler when get error tasks in multi-threads (#45268)
### Rationale for this change

When the TaskGroup should be canceled, it will move the number which not-start to finished to avoid do them(in `TaskSchedulerImpl::Abort`). But this is one operation that happens in multi-threads. At the same time, maybe some task start to running and happen some error. Then they will return the bad status.  
But the tasks are running for Scheduler, they will just return bad status and not change the running_task count. Because the code uses `RETURN_NOT_OK`.

### What changes are included in this PR?

For any task, what status weather it returns, it will change the running_count before return.

### Are these changes tested?

No. It is too hard to build ut.

### Are there any user-facing changes?

No. But I am very shocked at hasn't this happened to anyone?

* GitHub Issue: #45266

Lead-authored-by: zhouyunpei <zhouyunpei@yanhuangdata.com>
Co-authored-by: Rossi Sun <zanmato1984@gmail.com>
Co-authored-by: Antoine Pitrou <antoine@python.org>
Signed-off-by: Rossi Sun <zanmato1984@gmail.com>