GH-41816: [C++] Add Minimal Meson Build of libarrow (#45441)
### Rationale for this change

The Meson build system may be more user friendly to some developers, and may make it easier to perform tasks like valgrind, coverage, or ASAN/UBSAN coverage. There is also a prior art for using meson in the nanoarrow and arrow-adbc projects.

### What changes are included in this PR?

This PR implements a Meson configuration that can build a minimal libarrow.

### Are these changes tested?

A nightly CI job is also added to detect regressions.

### Are there any user-facing changes?

No

* GitHub Issue: #41816

Lead-authored-by: Will Ayd <william.ayd@icloud.com>
Co-authored-by: Sutou Kouhei <kou@clear-code.com>
Signed-off-by: Sutou Kouhei <kou@clear-code.com>