GH-43135: [R] Change the binary type mapping to `blob::blob` (#45595)
### Rationale for this change

Many packages, including `nanoarrow`, use the `blob` class provided by the `blob` package to represent a vector of binary data.
However, this package did not use `blob`.
Therefore, when this package convert Arrow Binary type to R, it was difficult for other packages to interpret it as a Binary type.

### What changes are included in this PR?

The current special processing for the `arrow_binary` class should also be performed for the `blob` class.
Also, `arrow_binary`, `arrow_large_binary`, and `arrow_fixed_size_binary` will be changed to subclasses of the `blob` class.

### Are these changes tested?

Rewrote some existing tests to use `blob` instead of `arrow_binary`.

### Are there any user-facing changes?

Yes.

But the traditional classes have been changed to subclasses of `blob`, so little impact is expected.
* GitHub Issue: #43135

Authored-by: SHIMA Tatsuya <ts1s1andn@gmail.com>
Signed-off-by: Jacob Wujciak-Jens <jacob@wujciak.de>