GH-45905: [C++][Acero] Enlarge the timeout in ConcurrentQueue test to reduce sporadical failures (#45923)
### Rationale for this change

The original test doesn't establish proper synchronization between `queue.WaitAndPop()` in one thread and the `fut_pop.get()` in the other. I.e., if the `std::async` thread (doing `queue.WaitAndPop()`) is scheduled by the os later than the timeout set for the `fut_pop.wait_for()` in the main thread, the timeout is returned and the assertion fires.

### What changes are included in this PR?

Enlarge the timeout from `10ms` to `5s` to reduce the chance of timing related failures.

Note in theory we should use `wait` (block forever until fulfilled) to prevent the timing issue, but in practice `5s` should be enough and it prevents the CI from hanging forever if any bug is introduced.

### Are these changes tested?

The fix is to the test.

### Are there any user-facing changes?

None.

* GitHub Issue: #45905

Authored-by: Rossi Sun <zanmato1984@gmail.com>
Signed-off-by: Rossi Sun <zanmato1984@gmail.com>