GH-41985: [Python][Docs] Clarify docstring of pyarrow.compute.scalar() (#45668)
### Rationale for this change

The current docstring for `pyarrow.compute.scalar()` is incomplete and potentially misleading. It states that it only accepts "a subset of types" when in reality it accepts any value that can be converted to a `pyarrow.Scalar`. Additionally, there's no clear explanation of the difference between `pyarrow.scalar()` and `pyarrow.compute.scalar()`, which can be confusing for users since both functions have the same name but serve different purposes.

### What changes are included in this PR?

This PR updates the docstring for `pyarrow.compute.scalar()` to clarify that it accepts any value convertible to a `pyarrow.Scalar` and add a detailed explanation of the difference between `pyarrow.scalar()` and `pyarrow.compute.scalar()`

### Are these changes tested?

via CI.

### Are there any user-facing changes?

No.

* GitHub Issue: #41985

Lead-authored-by: Chilin Chiou <chilin.chiou@gmail.com>
Co-authored-by: ChiLin Chiu <chilin.chiou@gmail.com>
Co-authored-by: Alenka Frim <AlenkaF@users.noreply.github.com>
Co-authored-by: Joris Van den Bossche <jorisvandenbossche@gmail.com>
Signed-off-by: AlenkaF <frim.alenka@gmail.com>